requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.7
coloredlogs==15.0.1
fastapi==0.103.2
filelock==3.20.1
flatbuffers==25.9.23
fsspec==2025.12.0
h11==0.14.0
hf-xet==1.2.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.20.0
onnx-ir==0.1.12
onnxruntime-gpu==1.23.2
onnxscript==0.5.6
packaging==25.0
pluggy==1.6.0
protobuf==6.33.2
pydantic==2.4.2
pydantic-core==2.10.1
pygments==2.19.2
pytest==7.4.2
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.31.0
rich==13.6.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typer-slim==0.20.0
typing-extensions==4.15.0
urllib3==2.6.2
uvicorn==0.23.2
